requirements.txt
flask==3.1.2
flask-cors==5.0.0
pymongo==4.11.2
python-dotenv==1.2.1
sentence-transformers==3.3.1
google-generativeai==0.8.5
gunicorn==21.2.0
